create "main_space" for all .textContent
Created a main_space <div> for all .textContent to be displayed.
Instead of having multiple event listeners for each button, I created an id #buttons that dealt with all buttons.
Then using an event listener for #buttons, with an if statement that would have main_space.textContent = to the .textContent of the targeted button.
Lastly, I made it so that clicking the DONE button no longer made all .textContents an empty string and instead made only the main_space.textContent an empty string, clearing the "main space" of the page. I did this because updating the individual textContents with the DONE button basically made all other buttons useless as their values were all set to an empty string. Doing this let each button keep their original values.