Add const for main_space
Added a const main_space_div that is equal to the main_space div.
Commented out the original listener for clicking the buttons and made a new one that utilizes .innerHTML to update/replace the content of main_space div.
The goal, and what seems to be happening is now I can actually add an id to the text being added and thus be able to hopefully click on the text that appears!