Add back button
Added a back to members button that updates the innerHTML of main_space_div.
I am now realizing that I now need to work with loops as I'm running into the same problem I first had with ifs when learning JS. Which was basically a neverending tree of options.
It's time to clean things up.